Update shared numpy strategy
